Version 1.3.7: Aims to be compatible with sdcc >= 4.2
sdcc 4.2.0 and later changes the calling convention for C functions.
Instead of passing all arguments on the stack, some can now be passed in
registers.

No problem if your whole program is written in C, but for flash4 which
is a mix of C and assembly we need to either rewrite the assembly
routines to use register arguments, or tell the compiler to use the
older calling convention for those functions. I chose the latter.

However, this still doesn't fix everything, because the standard library
is built with the expectation that some routines use the newer calling
convention, and the compiler flag to set the calling convention does not
affect the standard library (which was built without this flag). We
supply putchar() which the standard library expects to use the
compiler's native calling convention.

This commit changes the calling convention for all assembler routines to
stack arguments, and provides a version of putchar() that should build
with register or stack arguments according to the compiler's
expectations.